Add functionality to allow user to edit own credentials

+++ backend/controllers/updateUser.js
@@ -13,35 +13,40 @@ export default async (req, res) => {
       [password, password, email, email, username]
     );
 
-    // get current groups
-    const [joinedGroups, fields] = await connection.query(
-      `SELECT groupname from user_groups
+    //only admin can edit groups
+    if (req.permittedGroups && req.permittedGroups.contain("Admin")) {
+      // get current groups
+      const [joinedGroups, fields] = await connection.query(
+        `SELECT groupname from user_groups
       WHERE username = ?;`,
-      username
-    );
-    const currentJoinedGroups = joinedGroups.map((currentGroup) => currentGroup.groupname);
-    // remove unselected groups
-    const groupsToRemove = currentJoinedGroups.filter((curr_group) => !groups.includes(curr_group));
+        username
+      );
+      const currentJoinedGroups = joinedGroups.map((currentGroup) => currentGroup.groupname);
+      // remove unselected groups
+      const groupsToRemove = currentJoinedGroups.filter(
+        (curr_group) => !groups.includes(curr_group)
+      );
 
-    if (groupsToRemove.length > 0) {
-      const groupsPlaceholder = groupsToRemove.map(() => "?").join(", ");
-      await connection.query(
-        `DELETE FROM user_groups
+      if (groupsToRemove.length > 0) {
+        const groupsPlaceholder = groupsToRemove.map(() => "?").join(", ");
+        await connection.query(
+          `DELETE FROM user_groups
       WHERE username = '${username}' and
       groupname in (${groupsPlaceholder});`,
-        groupsToRemove
-      );
-    }
+          groupsToRemove
+        );
+      }
 
-    // assign user to groups selected
-    const groupsToAdd = groups.filter((group) => !currentJoinedGroups.includes(group));
-    if (groupsToAdd.length > 0) {
-      const newGroupsPlaceholder = groupsToAdd.map(() => "(?, ?)").join(", ");
-      const values = groupsToAdd.flatMap((group) => [group, username]);
-      await connection.query(
-        `INSERT INTO user_groups (groupname, username) VALUES ${newGroupsPlaceholder};`,
-        values
-      );
+      // assign user to groups selected
+      const groupsToAdd = groups.filter((group) => !currentJoinedGroups.includes(group));
+      if (groupsToAdd.length > 0) {
+        const newGroupsPlaceholder = groupsToAdd.map(() => "(?, ?)").join(", ");
+        const values = groupsToAdd.flatMap((group) => [group, username]);
+        await connection.query(
+          `INSERT INTO user_groups (groupname, username) VALUES ${newGroupsPlaceholder};`,
+          values
+        );
+      }
     }
 
     res.status(200).json({
